requirements.txt
Django~=5.0.3
django-cors-headers~=4.3.1
django-filter~=24.2
djangorestframework~=3.15.1
djangorestframework-simplejwt~=5.3.1
drf-yasg~=1.21.7
pillow~=10.3.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
setuptools~=69.5.1